firebase 토큰 전송, 모드/방향 메시지 전송

+++ app/src/main/java/kr/ac/pusan/walkover/autotrackingcctv/retrofit/Mode.java
@@ -4,10 +4,10 @@ package kr.ac.pusan.walkover.autotrackingcctv.retrofit;
  * Created by soeun on 2017-08-01.
  */
 public class Mode {
-    private Integer camera_num;
+    private long camera_num;
     private String mode;
 
-    public Mode(Integer camera_num, String mode) {
+    public Mode(long camera_num, String mode) {
         this.camera_num = camera_num;
         this.mode = mode;
     }
